pyproject.toml
[project]
name = "scrape-pinterest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bertopic>=0.17.4",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
    "selenium>=4.41.0",
    "umap-learn>=0.5.11",
    "webdriver-manager>=4.0.2",
]
